requirements.txt
google-adk>=1.0.0
pytest>=9.0.0
pytest-asyncio>=1.0.0
pandas>=2.0.0
rouge-score>=0.1.2
